# swagger.yaml
openapi: 3.0.0
info:
  title: sophs ecommerce api
  contact:
    name: Ebune Daniel Yahya
    email: ebunedy@gmail.com
  version: "1.0"
servers:
  - url: https://sophdev.herokuapp.com
    variables: {}
paths:
  /api/v1/users/login:
    post:
      tags:
        - Users
      summary: login user
      operationId: loginuser
      parameters: []
      requestBody:
        description: log a user in
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/loginrequest"
            example:
              email: example@gmail.com
              password: password
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
      servers:
        - url: https://sophdev.herokuapp.com
          variables: {}
  /api/v1/users/register:
    post:
      tags:
        - Users
      summary: register user
      operationId: registeruser
      parameters: []
      requestBody:
        description: create a user
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/registerrequest"
            example:
              firstName: ebune
              lastName: daniel
              email: example@gmail.com
              phoneNumber: "08122222222"
              password: password
              comfirmPassword: password
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
      servers:
        - url: https://sophdev.herokuapp.com
          variables: {}
  /api/v1/users:
    get:
      tags:
        - Users
      summary: get all users
      operationId: getallusers
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
      servers:
        - url: https://sophdev.herokuapp.com
          variables: {}
  /api/v1/users/single-user:
    get:
      tags:
        - Users
      summary: get single user
      operationId: getsingleuser
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /api/v1/products/uploads:
    post:
      tags:
        - Products
      summary: image upload
      operationId: imageupload
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
  /api/v1/products:
    post:
      tags:
        - Products
      summary: create product
      operationId: createproduct
      parameters: []
      requestBody:
        description: create a product
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createproductrequest"
            example:
              name: headset
              category: phone accessory
              description: a good headset
              specification: product specification
              featured: false
              price: 6
              images:
                - https://res.cloudinary.com/sophdev/image/upload/v1657188985/file-upload/tmp-1-1657188982781.jpg
                - https://res.cloudinary.com/sophdev/image/upload/v1657189727/file-upload/tmp-2-1657189723524.jpg
              company: our company
              colors:
                - cyan
                - red
              freeShipping: false
              inventory: 20
              size:
                - m
                - xl
              user: 62b731103f4fb7f974b
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      servers:
        - url: https://sophdev.herokuapp.com
          variables: {}
    get:
      tags:
        - Products
      summary: get all products
      operationId: getallproducts
      parameters:
        - name: category
          in: query
          description: sort products by category. e.g computer, women fashion
          style: form
          explode: true
          schema:
            type: string
            example: women fashion
        - name: sort
          in: query
          description: sort products by oldest or latest
          style: form
          explode: true
          schema:
            type: string
            example: latest
        - name: page
          in: query
          description: for pagination
          style: form
          explode: true
          schema:
            type: string
            example: "1"
        - name: search
          in: query
          description: search products by a word found in product name
          style: form
          explode: true
          schema:
            type: string
            example: men sneakers
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
      servers:
        - url: https://sophdev.herokuapp.com
          variables: {}
  /api/v1/products/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The product ID
    get:
      tags:
        - Products
      summary: get single product
      operationId: getsingleproduct
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      servers:
        - url: https://sophdev.herokuapp.com
          variables: {}
    patch:
      tags:
        - Products
      summary: update product
      operationId: updateproduct
      parameters: []
      requestBody:
        description: update product
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/updateproductrequest"
            example:
              name: headset
              category: phone accessory
              description: excellent headset
              images:
                - https://res.cloudinary.com/sophdev/image/upload/v1657188985/file-upload/tmp-1-1657188982781.jpg
                - https://res.cloudinary.com/sophdev/image/upload/v1657189727/file-upload/tmp-2-1657189723524.jpg
              company: our company
              colors:
                - green
                - blue
              user: 62b731103f4fb7f974be4bec
              price: 6
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      servers:
        - url: https://sophdev.herokuapp.com
          variables: {}
    delete:
      tags:
        - Products
      summary: delete-product
      operationId: delete-product
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      servers:
        - url: https://sophdev.herokuapp.com
          variables: {}
  /api/v1/reviews:
    post:
      tags:
        - reviews
      summary: create review
      operationId: createreview
      parameters: []
      requestBody:
        description: review creation
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createreviewrequest"
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      servers:
        - url: https://sophdev.herokuapp.com
          variables: {}
    get:
      tags:
        - reviews
      summary: get all reviews
      operationId: getallreviews
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      security: []
      servers:
        - url: https://sophdev.herokuapp.com
          variables: {}
  /api/v1/reviews/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The review ID
    get:
      tags:
        - reviews
      summary: get single review
      operationId: getsinglereview
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      servers:
        - url: https://sophdev.herokuapp.com
          variables: {}
    patch:
      tags:
        - reviews
      summary: update review
      operationId: updatereview
      parameters: []
      requestBody:
        description: ""
        content:
          text/plain:
            schema:
              type: string
              example:
                title: great product
                rating: 4
                comment: A great product, worth every penny I paid for
            example:
              title: great product
              rating: 4
              comment: A great product, worth every penny I paid for
        required: true
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      servers:
        - url: https://sophdev.herokuapp.com
          variables: {}
    delete:
      tags:
        - reviews
      summary: delete review
      operationId: deletereview
      parameters: []
      responses:
        "200":
          description: ""
          headers: {}
      deprecated: false
      servers:
        - url: https://sophdev.herokuapp.com
          variables: {}
components:
  schemas:
    loginrequest:
      title: loginrequest
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      description: ""
      example:
        email: example@gmail.com
        password: password
    registerrequest:
      title: registerrequest
      required:
        - firstName
        - lastName
        - email
        - phoneNumber
        - password
        - comfirmPassword
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        password:
          type: string
        comfirmPassword:
          type: string
      example:
        firstName: ebune
        lastName: daniel
        email: example@gmail.com
        phoneNumber: "08120776546"
        password: password
        comfirmPassword: pasword
    createproductrequest:
      title: createproductrequest
      required:
        - name
        - category
        - description
        - images
        - company
        - colors
        - user
        - price
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
          description: ""
        company:
          type: string
        colors:
          type: array
          items:
            type: string
          description: ""
        user:
          type: string
        price:
          type: integer
          format: int32
      example:
        name: headset
        category: phone accessory
        description: a good headset
        specification: product specification
        featured: false
        price: 6
        images:
          - https://res.cloudinary.com/sophdev/image/upload/v1657188985/file-upload/tmp-1-1657188982781.jpg
          - https://res.cloudinary.com/sophdev/image/upload/v1657189727/file-upload/tmp-2-1657189723524.jpg
        company: our company
        colors:
          - cyan
          - red
        freeShipping: false
        inventory: 20
        size:
          - m
          - xl
        user: 62b731103f4fb7f974b
    updateproductrequest:
      title: updateproductrequest
      required:
        - name
        - category
        - description
        - images
        - company
        - colors
        - user
        - price
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
          description: ""
        company:
          type: string
        colors:
          type: array
          items:
            type: string
          description: ""
        user:
          type: string
        price:
          type: integer
          format: int32
      example:
        name: headset
        category: funiture
        description: a goood furniture
        images:
          - https://res.cloudinary.com/sophdev/image/upload/v1657188985/file-upload/tmp-1-1657188982781.jpg
          - https://res.cloudinary.com/sophdev/image/upload/v1657189727/file-upload/tmp-2-1657189723524.jpg
        company: our company
        colors:
          - green
          - blue
        user: 62b731103f4fb7f974be4bec
        price: 6
    createreviewrequest:
      title: createreviewrequest
      required:
        - title
        - rating
        - comment
        - user
        - product
      type: object
      properties:
        title:
          type: string
        rating:
          type: integer
          format: int32
        comment:
          type: string
        user:
          type: string
        product:
          type: string
      example:
        title: great product
        rating: 4
        comment: A great product, worth every penny I paid for
        user: 62cc00fc48d0579a3ecb7b76
        product: 62c8038c551f16bfb1e199b0
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
  - httpBearer: []
tags:
  - name: Users
  - name: Products
  - name: reviews